+++ metautil.d.ts
@@ -72,9 +72,11 @@ export interface QueueElement {
 
 export class Semaphore {
   constructor(concurrency: number, size?: number, timeout?: number);
+  concurrency: number;
   counter: number;
   timeout: number;
   size: number;
+  empty: boolean;
   queue: Array<QueueElement>;
   enter(): Promise<void>;
   leave(): void;